Encrypt Password
A method to encrypt password has been added. How it works so far is, the password in the database is already encrypted, but to make them equal, you have to encrypt the password input, so when you log in, the system will read the password as an encryption. This method can in the process of making the system be used to encrypt password in the future then add them into the database.